refactor: add static metadata object to root layout
