+++ app/(app)/players/page.tsx
@@ -2,7 +2,7 @@
 
 import Link from 'next/link';
 import { useEffect } from 'react';
-import { usePlayerService } from '../../services';
+import { usePlayerService } from '../../../services';
 
 export default function Players() {
     const playerService = usePlayerService();
