+++ app/layout.tsx
@@ -1,35 +1,25 @@
-"use client"
+import type { Metadata } from 'next'
+import { Inter } from 'next/font/google'
+import './globals.css'
 
-import 'globals.css';
-import { Sidebar } from '../components/sidebar';
-import { useState } from 'react';
-import classNames from 'classnames';
+const inter = Inter({ subsets: ['latin'] })
 
-export default function Rootlayout({
-  children
+export const metadata: Metadata = {
+  title: {
+    template: '%s | FP Manager',
+    default: 'FP Manager',
+  },
+  description: 'Manage video settings',
+}
+
+export default function RootLayout({
+  children,
 }: {
   children: React.ReactNode
 }) {
-
-  const [collapsed, setSidebarCollapsed] = useState(false);
-
   return (
     <html lang="en">
-      <body>
-        <div
-          className={classNames({
-            "lg:grid min-h-screen": true,
-            "lg:grid-cols-app": !collapsed,
-            "lg:grid-cols-collapsed": collapsed,
-            "transition-[grid-template-columns] duration-300 ease-in-out": true,
-          })}
-        >
-          <Sidebar setSidebarCollapsed={setSidebarCollapsed} isCollapsed={collapsed} />
-          <main className="max-w-screen px-4 pb-12 pt-24 space-y-2 lg:col-start-2 lg:w-auto lg:px-6 lg:pt-8">
-            {children}
-          </main>
-        </div>
-      </body>
+      <body className={inter.className}>{children}</body>
     </html>
-  );
+  )
 }
